requirements.txt
streamlit
pandas
matplotlib
requests
paho-mqtt